Test Structure Overhaul SM Complete
