+++ tests/SMTests/smVerifier.py
